requirements.txt
boto==2.42.0
dj-database-url==0.4.0
Django==1.9.2
django-jsonify==0.3.0
django-storages==1.5.0
django-tinymce==2.3.0
gunicorn==19.4.5
psycopg2==2.6.1
sorl-thumbnail==12.3
whitenoise==2.0.6
